+++ src/main/java/mod/render360/coretransform/classtransformers/GuiScreenTransformer.java
@@ -15,7 +15,7 @@ import mod.render360.coretransform.RenderUtil;
 import mod.render360.coretransform.classtransformers.name.ClassName;
 import mod.render360.coretransform.classtransformers.name.MethodName;
 import mod.render360.coretransform.classtransformers.name.Names;
-import mod.render360.coretransform.render.RenderMethod;
+import mod.render360.render.RenderMethod;
 
 import static org.objectweb.asm.Opcodes.*;
