+++ src/main/java/mod/render360/gui/Render360Settings.java
@@ -1,4 +1,4 @@
-package mod.render360.coretransform.gui;
+package mod.render360.gui;
 
 import java.io.File;
 import java.io.IOException;
@@ -7,7 +7,7 @@ import java.nio.file.Paths;
 import java.util.List;
 
 import mod.render360.coretransform.RenderUtil;
-import mod.render360.coretransform.render.RenderMethod;
+import mod.render360.render.RenderMethod;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.resources.I18n;
