+++ src/main/java/mod/render360/render/Equirectangular.java
@@ -1,11 +1,11 @@
-package mod.render360.coretransform.render;
+package mod.render360.render;
 
 import java.util.List;
 
 import org.lwjgl.opengl.GL20;
 
-import mod.render360.coretransform.Shader;
-import mod.render360.coretransform.gui.Slider;
+import mod.render360.Shader;
+import mod.render360.gui.Slider;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiPageButtonList.GuiResponder;
