+++ src/main/java/mod/render360/render/RenderMethod.java
@@ -1,4 +1,4 @@
-package mod.render360.coretransform.render;
+package mod.render360.render;
 
 import java.io.IOException;
 import java.util.List;
@@ -9,10 +9,10 @@ import org.lwjgl.opengl.GL13;
 import org.lwjgl.opengl.GL15;
 import org.lwjgl.opengl.GL20;
 
+import mod.render360.Shader;
 import mod.render360.coretransform.CLTLog;
 import mod.render360.coretransform.RenderUtil;
-import mod.render360.coretransform.Shader;
-import mod.render360.coretransform.gui.Slider;
+import mod.render360.gui.Slider;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiPageButtonList.GuiResponder;
