+++ src/main/java/mod/render360/render/Standard.java
@@ -1,9 +1,9 @@
-package mod.render360.coretransform.render;
+package mod.render360.render;
 
 import java.util.List;
 
+import mod.render360.Shader;
 import mod.render360.coretransform.RenderUtil;
-import mod.render360.coretransform.Shader;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.renderer.EntityRenderer;
